refactor: move to constants
