+++ src/app/api/reactions/[slug]/route.ts
@@ -1,8 +1,8 @@
 import type { ReactionType } from '@prisma/client'
 import type { NextRequest } from 'next/server'
 
+import { MAX_REACTIONS_PER_SESSION } from '@/constants/reactions'
 import { getErrorMessage, response } from '@/lib/api'
-import { MAX_REACTIONS_PER_SESSION } from '@/lib/constants'
 import { getSessionId } from '@/lib/server'
 import { createReaction, getReactions } from '@/services/reactions'
